Add logsearchapi
